Safari-609.1.11.1/LayoutTests/loader/navigation-policy/should-open-external-urls/subframe-navigated-programatically-by-main-frame.html

<script>

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpPolicyDelegateCallbacks();
    testRunner.queueLoad("resources/main-frame-with-subframe-main-programatically-navigates-subframe.html", "", true);
}

</script>
This page has does not have the "should open external URLs" flag set.<br>
It navigates to a 2nd page that does have the flag set.<br>
That 2nd page contains an iframe, and the main frame programatically navigates the iframe to a 3rd document.<br>
The "should open external URLs" flag should be set because the programatic navigation initiated from the main frame and the flag should propagate.<br>
