LayoutTests/loader/navigation-policy/should-open-external-urls/subframe-navigated-programatically-by-main-frame-expected.txt

 - decidePolicyForNavigationAction
<NSURLRequest URL resources/main-frame-with-subframe-main-programatically-navigates-subframe.html, main document URL resources/main-frame-with-subframe-main-programatically-navigates-subframe.html, http method GET> is main frame - yes should open URLs externally - yes
 - decidePolicyForNavigationAction
<NSURLRequest URL do-nothing.html, main document URL main-frame-with-subframe-main-programatically-navigates-subframe.html, http method GET> is main frame - no should open URLs externally - no
 - decidePolicyForNavigationAction
<NSURLRequest URL notify-done.html, main document URL main-frame-with-subframe-main-programatically-navigates-subframe.html, http method GET> is main frame - no should open URLs externally - yes
